+++ var/starter_kits/italiagov/webpack.prod.js
@@ -1,10 +1,12 @@
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const { CleanWebpackPlugin } = require('clean-webpack-plugin')
-const { merge } = require('webpack-merge')
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import { CleanWebpackPlugin } from 'clean-webpack-plugin';
+import { merge } from 'webpack-merge';
+import TerserPlugin from 'terser-webpack-plugin';
+import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 
-const common = require('./webpack.common')
+import common from './webpack.common.js';
 
-module.exports = merge(common, {
+export default merge(common, {
   mode: 'production',
   devtool: false,
 
@@ -12,6 +14,7 @@ module.exports = merge(common, {
     // Removes/cleans build folders and unused assets when rebuilding
     new CleanWebpackPlugin(),
   ],
+
   module: {
     rules: [
       {
@@ -24,7 +27,7 @@ module.exports = merge(common, {
               importLoaders: 2,
               sourceMap: false,
               modules: false,
-            }
+            },
           },
           'postcss-loader',
           'sass-loader',
@@ -32,12 +35,39 @@ module.exports = merge(common, {
       },
     ],
   },
+
   optimization: {
     minimize: true,
+    minimizer: [
+      // Aggiungi TerserPlugin per minimizzare il JavaScript
+      new TerserPlugin({
+        terserOptions: {
+          compress: {
+            drop_console: true, // Rimuove i console.log dal codice di produzione
+          },
+          format: {
+            comments: false, // Rimuove tutti i commenti
+          },
+        },
+        extractComments: false, // Impedisce l'estrazione dei commenti in file separati
+      }),
+      // Aggiungi CssMinimizerPlugin per minimizzare il CSS
+      new CssMinimizerPlugin({
+        minimizerOptions: {
+          preset: [
+            'default',
+            {
+              svgo: false, // Disabilita SVGO
+            },
+          ],
+        },
+      }),
+    ],
   },
+
   performance: {
     hints: false,
     maxEntrypointSize: 860000,
     maxAssetSize: 860000,
   },
-})
+});